+++ src/commit/pre-commit-shell.js
@@ -2,8 +2,8 @@
  * @Github: https://github.com/OBKoro1
  * @Author: OBKoro1
  * @Created_time: 2019-08-31 15:01:52
- * @LastEditors: Koro
- * @LastEditTime: 2019-09-02 21:07:18
+ * LastEditors: OBKoro1
+ * LastEditTime: 2019-09-04 19:31:52
  * @Description: 当没有commit文件时，默认创建的pre-commit 文件
  */